efuse: fixed a stack-out-of-bounds issue. [1/1]
PD#SWPL-79503

Problem:
With kasan enable,the stack is out of bounds when read efuse key.

Solution:
Modify the type of variable 'retcnt' from int to long.

Verify:
T982-x9

+++ drivers/amlogic/efuse_unifykey/efuse_api.c
@@ -365,7 +365,7 @@ static ssize_t _efuse_read(char *buf, size_t count, loff_t *ppos)
 	unsigned int pos = *ppos;
 
 	struct efuse_hal_api_arg arg;
-	unsigned int retcnt;
+	unsigned long retcnt;
 	ssize_t ret;
 
 	arg.cmd = EFUSE_HAL_API_READ;
@@ -387,7 +387,7 @@ static ssize_t _efuse_write(const char *buf, size_t count, loff_t *ppos)
 	unsigned int pos = *ppos;
 
 	struct efuse_hal_api_arg arg;
-	unsigned int retcnt;
+	unsigned long retcnt;
 	ssize_t ret;
 
 	arg.cmd = EFUSE_HAL_API_WRITE;
